add option --directory/-d
If platsch is called as PID 1 in early userspace it is given the kernel's
commandline and uses a hardcoded directory to look for the image to display. As
it can potentially be called multiple times from later userspace one probably
wants to direct it to different directories with different images.

Acked-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>
Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>