add option --basename/-b
The filename of the image to show is constructed from a basename by adding the
display's width & height as well as the specifier of the format that the raw
image has been saved in and finally the ending ".bin". Enable the user to choose
the basename by adding a corresponding option.

Acked-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>
Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>