allow directory/basename selection via environment
Introduce the possibility to choose the directory searched for the splash-
images, as well as the images' basename via the process' environment

  platsch_directory=/usr/share/plitschplatsch
  platsch_basename=splosh

with the corresponding commandline parameters still having higher priority
compared to the environment.

The Kernel passes unrecognized key-value parameters not containing dots into
init’s environment [1]. So the above environment variable can be supplied via
the kernel commandline. This also allows dynamic use cases where the bootloader
decides which resolution to use on which connector.

[1] https://www.kernel.org/doc/html/latest/admin-guide/kernel-parameters.html

Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>