platsch: fix crtc_id usage in drmprepare_crtc
The crtc_id is used as crtc_id but also as indicator that the crtc_id is
already in use. For this crtc_id was declared as int32. This cause a
sign compare issue when compared to unsigned int.

Introduce a local boolean to check if the crtc_id is already in use and
declare the crtc_id as uint32_t to fix this.

Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>