libplatsch: add support for custom draw operations
This adds the support to register a custom draw operation callback which
will be called during platsch_draw(). The callback has all the necessary
information about the framebuffer and a reference to the framebuffer
itself to update content. The information will be freed after the
callback has finished, therfore the content needs to be copied by the
application.

Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>