libplatsch: add proper API export macro
At the moment we rely on the 'default' visibility which was okay since
all global functions are API functions. This is error prone as one could
expose symbols by accident.

Therefore change the behavior by setting the default visibility to
'hidden' and introducing a new macro to export selected API functions.

Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>