libplatsch: handle possible clones
At the moment we miss the support to drive multiple connectors which are
connected to the same CRTC, also called a hardware clone mode.

libplatsch is not using the DRM atomic API due to it limited scope
therefore using properties values isn't sufficient. Instead all possible
connector ids need to be passed during the drmModeSetCrtc().

To find all connectors which belong to the same CRTC, the encoder behind
each connector must be validated by making use of the encoders
'.possible_clones' array.

If a connector shall be added which belongs to an already used CRTC, a check
is performed to see if this connector can be driven by the CRTC. In that
case the connector id is added to the 'struct modeset_dev::conn_id'
array and -EEXIST is returned.

Because of this error, no new 'struct modeset_dev' is added and certain
error() messages aren't printed because this can be a valid use-case.

Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>